+++ Product.cs
@@ -6,10 +6,10 @@ namespace E_CommerceApplication
         public string ProductID { get; set; }
         public string ProductName { get; set; }
         public int Stock { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public int ShippingDuration { get; set; }
 
-        public Product(string productName, int stock, int price, int shippingDuration)
+        public Product(string productName, int stock, decimal price, int shippingDuration)
         {
             ProductID = $"PID{s_productID++}";
             ProductName = productName;
